+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to Sparvi will be documented in this file.
 
+## [0.3.0] - 2025-02-28
+### Added
+- Support for Supabase logins
+
+### Changed
+- Updated login on back and front to use Supabase
+
+### Fixed
+- None
+
 ## [0.2.0] - 2025-02-27
 ### Added
 - Validation rules feature for defining custom data quality checks
@@ -12,8 +22,20 @@ All notable changes to Sparvi will be documented in this file.
 - Sample validation rules for common data quality checks
 - Documentation section in header
 
+### Changed
+- None
+
+### Fixed
+- None
+
 ## [0.1.0] - 2025-02-26
 ### Added
 - Initial implementation of data profiler
 - Basic dashboard UI
 - Support for DuckDB connections
+
+### Changed
+- None
+
+### Fixed
+- None